Load the currently playing widget in with the rest of the home row
